requirements.txt
atomicwrites==1.4.0
attrs==20.3.0
colorama==0.4.4
coverage==5.5
Django==2.2.7
django-filter==2.4.0
djangorestframework==3.12.2
Faker==6.5.0
iniconfig==1.1.1
Markdown==3.3.4
mixer==7.1.2
packaging==20.9
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.2
pytest-cov==2.11.1
pytest-django==4.1.0
pytest-html==3.1.1
pytest-metadata==1.11.0
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
